% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapiR_connection.R
\name{authenticate_osmapi}
\alias{authenticate_osmapi}
\alias{logout_osmapi}
\title{Authenticate or logout osmapiR}
\usage{
authenticate_osmapi()

logout_osmapi()
}
\value{
For \code{authenticate_osmapi}, print the user and permissions of the connection and return invisibly the display
name of the logged user. \code{logout_osmapi} clear the OAuth2 token and can be useful to change user.
}
\description{
Log in/out osmapiR.
}
\details{
All functions that require authentication will trigger the log in if the session is not yet authenticated, so calling
this function is not really needed. Use \code{authenticate_osmapi} to sign in before executing scripts that require
authentication to avoid interruptions.
}
\examples{
\dontrun{
authenticate_osmapi()
logout_osmapi()
}
}
\seealso{
Other API functions: 
\code{\link{osm_api_versions}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_permissions}()},
\code{\link{set_osmapi_connection}()}
}
\concept{API functions}
