% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_osmapi_url}
\alias{get_osmapi_url}
\title{get_osmapi_url}
\usage{
get_osmapi_url()
}
\value{
The OSM API URL
}
\description{
Return the URL of the specified OSM API. Default is \url{https://api.openstreetmap.org/}.
}
\seealso{
\code{\link[=set_osmapi_url]{set_osmapi_url()}}

Other osmapi: 
\code{\link{set_osmapi_url}()}
}
\concept{osmapi}
