% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{osm_bbox_objects}
\alias{osm_bbox_objects}
\title{Retrieve map data by bounding box}
\usage{
osm_bbox_objects(bbox)
}
\arguments{
\item{bbox}{Coordinates for the area to retrieve the map data from (\verb{left,bottom,right,top}).}
}
\description{
The following command returns:
\itemize{
\item All nodes that are inside a given bounding box and any relations that reference them.
\item All ways that reference at least one node that is inside a given bounding box, any relations that reference them
[the ways], and any nodes outside the bounding box that the ways may reference.
\item All relations that reference one of the nodes, ways or relations included due to the above rules. (Does '''not'''
apply recursively, see explanation below.)
}
}
\details{
Note that, while this command returns those relations that reference the aforementioned nodes and ways, the reverse
is not true: it does not (necessarily) return all of the nodes and ways that are referenced by these relations. This
prevents unreasonably-large result sets. For example, imagine the case where:
\itemize{
\item There is a relation named "England" that references every node in England.
\item The nodes, ways, and relations are retrieved for a bounding box that covers a small portion of England.
While the result would include the nodes, ways, and relations as specified by the rules for the command, including
the "England" relation, it would (fortuitously) \strong{not} include \strong{every} node and way in England. If desired, the
nodes and ways referenced by the "England" relation could be retrieved by their respective IDs.
}

Also note that ways which intersect the bounding box but have no nodes within the bounding box will not be returned.
}
\note{
For downloading data for purposes other than editing or exploring the history of the objects, perhaps is better to
use the Overpass API. A similar function to download OSM objects using Overpass, is implemented in \code{\link[osmdata:opq]{osmdata::opq()}}.
}
\examples{
map_data <- osm_bbox_objects(bbox = c(3.2164192, 42.0389667, 3.2317829, 42.0547099))
map_data
}
\seealso{
Other OSM objects' functions: 
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_history_object}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{OSM objects' functions}
