% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeset_discussion.R
\name{osm_comment_changeset_discussion}
\alias{osm_comment_changeset_discussion}
\title{Comment a changeset}
\usage{
osm_comment_changeset_discussion(changeset_id, comment)
}
\arguments{
\item{changeset_id}{The id of the changeset to comment represented by a numeric or a character value.}

\item{comment}{The text of the comment to post.}
}
\value{
Returns the changeset information invisibly as a \code{xml_document}.
}
\description{
Add a comment to a changeset and subscribe to the discussion. The changeset must be closed. Requires authentication.
}
\seealso{
Other changeset discussion's functions: 
\code{\link{osm_subscribe_changeset_discussion}()},
\code{\link{osm_unsubscribe_changeset_discussion}()}
}
\concept{changeset discussion's functions}
