% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_map_notes.R
\name{osm_create_comment_note}
\alias{osm_create_comment_note}
\title{Create a new comment in a note}
\usage{
osm_create_comment_note(note_id, text)
}
\arguments{
\item{note_id}{Note id represented by a numeric or a character value.}

\item{text}{The comment as arbitrary text.}
}
\description{
Add a new comment to an existeing note. Requires authentication.
}
\seealso{
Other edit notes' functions: 
\code{\link{osm_close_note}()},
\code{\link{osm_create_note}()},
\code{\link{osm_delete_note}()}
}
\concept{edit notes' functions}
