% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_traces.R
\name{osm_delete_gpx}
\alias{osm_delete_gpx}
\title{Delete GPS trace}
\usage{
osm_delete_gpx(gpx_id)
}
\arguments{
\item{gpx_id}{The track id represented by a numeric or a character value.}
}
\description{
Use this to delete a GPX file. Only usable by the owner account. Requires authentication.
}
