% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_delete_note}
\alias{osm_delete_note}
\title{Delete a note}
\usage{
osm_delete_note(note_id)
}
\arguments{
\item{note_id}{Note id represented by a numeric or a character value.}
}
\description{
Hide (delete) a note. This request needs to be done as an authenticated user with moderator role.
}
\details{
Use \code{\link[=osm_reopen_note]{osm_reopen_note()}}to make the note visible again.
}
\seealso{
Other functions for moderators: 
\code{\link{osm_hide_comment_changeset_discussion}()},
\code{\link{osm_redaction_object}()},
\code{\link{osm_unhide_comment_changeset_discussion}()}
}
\concept{functions for moderators}
