% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data.R
\name{osm_details_logged_user}
\alias{osm_details_logged_user}
\title{Details of the logged-in user}
\usage{
osm_details_logged_user()
}
\value{
A list with the user details.
}
\description{
You can get the home location, the display name of the user and other details.
}
\examples{
\dontrun{
usr_details <- osm_details_logged_user()
usr_details
}
}
\seealso{
Other users' functions: 
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{users' functions}
