% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_user_data.R
\name{osm_details_user}
\alias{osm_details_user}
\title{Details of a user}
\usage{
osm_details_user(user_id, format = c("R", "xml", "json"))
}
\arguments{
\item{user_id}{The id of the user to retrieve represented by a numeric or a character value (not the display name).}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}
}
\description{
Details of a user
}
\examples{
\dontrun{
usr <- osm_details_user(user_id = "11725140")
usr
}
}
\seealso{
Other users' functions: 
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_get_preferences_user}()}
}
\concept{users' functions}
