% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_user_data.R
\name{osm_details_users}
\alias{osm_details_users}
\title{Details of multiple users}
\usage{
osm_details_users(user_ids, format = c("R", "xml", "json"))
}
\arguments{
\item{user_ids}{The ids of the users to retrieve represented by a numeric or a character value (not the display
names).}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}
}
\description{
Details of multiple users
}
\examples{
\dontrun{
usrs <- osm_details_users(user_ids = c(1, 24, 44, 45, 46, 48, 49, 50))
usrs
}
}
\seealso{
Other users' functions: 
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_user}()},
\code{\link{osm_get_preferences_user}()}
}
\concept{users' functions}
