% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changesets.R
\name{osm_diff_upload_changeset}
\alias{osm_diff_upload_changeset}
\title{Diff (OsmChange format) upload to a changeset}
\usage{
osm_diff_upload_changeset(changeset_id, osmcha)
}
\arguments{
\item{changeset_id}{The ID of the changeset this diff belongs to. The user issuing this API call has to be the same
that created the changeset.}

\item{osmcha}{The OsmChange data. Can be the path of an OsmChange file, a \code{xml_document} or a data.frame inheriting
or following the structure of an \code{osmapi_OsmChange} object.}
}
\description{
With this API call files in the \href{https://wiki.openstreetmap.org/wiki/OsmChange}{OsmChange} format can be uploaded to
the server. This is guaranteed to be running in a transaction. So either all the changes are applied or none.
}
\note{
\itemize{
\item Processing stops at the first error, so if there are multiple conflicts in one diff upload, only the first
problem is reported.
\item Refer to \code{\link[=osm_capabilities]{osm_capabilities()}} --> \code{changesets$maximum_elements} for the maximum number of changes permitted in a
changeset.
\item There is currently no limit in the diff size on the Rails port. CGImap limits diff size to 50MB (uncompressed
size).
\item Forward referencing of placeholder ids is not permitted and will be rejected by the API.
}
}
\seealso{
Other edit changeset's functions: 
\code{\link{osm_create_changeset}()}
}
\concept{edit changeset's functions}
