% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changesets.R
\name{osm_download_changeset}
\alias{osm_download_changeset}
\title{Download a changeset}
\usage{
osm_download_changeset(changeset_id)
}
\arguments{
\item{changeset_id}{The id of the changeset represented by a numeric or a character value for which the OsmChange is
requested.}
}
\description{
Returns the \href{https://wiki.openstreetmap.org/wiki/OsmChange}{OsmChange} document describing all changes associated with the changeset.
}
\details{
\itemize{
\item The result of calling this may change as long as the changeset is open.
\item The elements in the OsmChange are sorted by timestamp and version number.
\item There is \code{\link[=osm_read_changeset]{osm_read_changeset()}} to get only information about the changeset itself.
}
}
\examples{
\dontrun{
chaset <- osm_download_changeset(changeset_id = 137003062)
chaset
}
}
\seealso{
Other get changesets' functions: 
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_changeset}()}
}
\concept{get changesets' functions}
