% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_feed_notes}
\alias{osm_feed_notes}
\title{RSS Feed of notes in a bbox}
\usage{
osm_feed_notes(bbox)
}
\arguments{
\item{bbox}{Coordinates for the area to retrieve the notes from (\verb{left,bottom,right,top}). Floating point numbers in
degrees, expressing a valid bounding box.}
}
\description{
RSS Feed of notes in a bbox
}
\examples{
feed_notes <- osm_feed_notes(bbox = c(0.8205414, 40.6686604, 0.8857727, 40.7493377))
## bbox as a character value also works. Equivalent call:
# feed_notes <- osm_feed_notes(bbox = "0.8205414,40.6686604,0.8857727,40.7493377")
feed_notes
}
\seealso{
Other get notes' functions: 
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_note}()},
\code{\link{osm_search_notes}()}
}
\concept{get notes' functions}
