% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{osm_fetch_objects}
\alias{osm_fetch_objects}
\title{Fetch objects}
\usage{
osm_fetch_objects(
  osm_type = c("nodes", "ways", "relations"),
  osm_ids,
  versions,
  format = c("R", "xml", "json"),
  tags_in_columns = FALSE
)
}
\arguments{
\item{osm_type}{Type of the objects(\code{"nodes"}, \code{"ways"} or \code{"relations"}).}

\item{osm_ids}{Object ids represented by a numeric or a character vector.}

\item{versions}{Version numbers for each object may be optionally provided.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}

\item{tags_in_columns}{if \code{FALSE} (default), the tags of the objects are saved in a single list column \code{tags}
containing a \code{data.frame} for each OSM object with the keys and values. If \code{TRUE}, add a column for each key.
Ignored if \code{format != "R"}.}
}
\description{
Fetch multiple objects of the same type at once.
}
\note{
For downloading data for purposes other than editing or exploring the history of the objects, perhaps is better to
use the Overpass API. A similar function to download OSM objects by \code{type} and \code{id} using Overpass, is implemented in
\code{\link[osmdata:opq_osm_id]{osmdata::opq_osm_id()}}.
}
\examples{
\dontrun{
node <- osm_fetch_objects(osm_type = "nodes", osm_ids = c(35308286, 1935675367))
node

way <- osm_fetch_objects(osm_type = "ways", osm_ids = c(13073736L, 235744929L))
way

# Specific versions
rel <- osm_fetch_objects(osm_type = "relations", osm_ids = c("40581", "341530"), versions = c(3, 1))
rel
}
}
\seealso{
Other OSM objects' functions: 
\code{\link{osm_bbox_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_history_object}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{OSM objects' functions}
