% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{osm_full_object}
\alias{osm_full_object}
\title{Full object}
\usage{
osm_full_object(
  osm_type = c("way", "relation"),
  osm_id,
  format = c("R", "xml", "json"),
  tags_in_columns = FALSE
)
}
\arguments{
\item{osm_type}{Object type (\code{"way"} or \code{"relation"}).}

\item{osm_id}{Object id represented by a numeric or a character value.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}

\item{tags_in_columns}{if \code{FALSE} (default), the tags of the objects are saved in a single list column \code{tags}
containing a \code{data.frame} for each OSM object with the keys and values. If \code{TRUE}, add a column for each key.
Ignored if \code{format != "R"}.}
}
\description{
This API call retrieves a way or relation and all other objects referenced by it.
}
\details{
For a way, it will return the way specified plus all nodes referenced by the way.
For a relation, it will return the following:
\itemize{
\item The relation itself
\item All nodes, ways, and relations that are members of the relation
\item Plus all nodes used by ways from the previous step
\item The same recursive logic is not applied to relations. This means: If relation r1 contains way w1 and relation r2,
and w1 contains nodes n1 and n2, and r2 contains node n3, then a "full" request for r1 will give you r1, r2, w1,
n1, and n2. Not n3.
}
}
\note{
For downloading data for purposes other than editing or exploring the history of the objects, perhaps is better to
use the Overpass API. A similar function to download OSM objects by \code{type} and \code{id} using Overpass, is implemented in
\code{\link[osmdata:opq_osm_id]{osmdata::opq_osm_id()}}.
}
\examples{
\dontrun{
way <- osm_full_object(osm_type = "way", osm_id = 13073736)
way

rel <- osm_full_object(osm_type = "relation", osm_id = "40581")
rel
}
}
\seealso{
Other OSM objects' functions: 
\code{\link{osm_bbox_objects}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_history_object}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_get_preferences_user}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{OSM objects' functions}
