% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_gps_traces.R
\name{osm_get_metadata_gpx}
\alias{osm_get_metadata_gpx}
\title{Download GPS Track Metadata}
\usage{
osm_get_metadata_gpx(gpx_id)
}
\arguments{
\item{gpx_id}{The track id represented by a numeric or a character value.}
}
\description{
Use this to access the metadata about a GPX file. Available without authentication if the file is marked public.
Otherwise only usable by the owner account and requires authentication.
}
\examples{
\dontrun{
trk_meta <- osm_get_metadata_gpx(gpx_id = 3498170)
trk_meta
}
}
\seealso{
Other get GPS' functions: 
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_list_gpxs}()}
}
\concept{get GPS' functions}
