% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_get_notes.R
\name{osm_get_notes}
\alias{osm_get_notes}
\title{Get notes}
\usage{
osm_get_notes(note_id, format = c("R", "xml", "rss", "json", "gpx"))
}
\arguments{
\item{note_id}{Note id represented by a numeric or a character value.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, \code{rss}, \code{json} or \code{gpx}.}
}
\value{
If \code{format = "R"}, returns a data frame with one map note per row. If \code{format = "json"}, returns a list with the json
structure. For \code{format} in \code{xml}, \code{rss}, and \code{gpx}, a \link[xml2:oldclass]{xml2::xml_document} with the corresponding format.
}
\description{
Returns the existing note with the given ID.
}
\examples{
\dontrun{
note <- osm_get_notes(note_id = "2067786")
note
}
}
\seealso{
Other get notes' functions: 
\code{\link{osm_feed_notes}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_search_notes}()}
}
\concept{get notes' functions}
