% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_traces.R
\name{osm_get_points_gps}
\alias{osm_get_points_gps}
\title{Get GPS Points}
\usage{
osm_get_points_gps(bbox, page_number = 0)
}
\arguments{
\item{bbox}{Coordinates for the area to retrieve the notes from (\verb{left,bottom,right,top}). Floating point numbers in
degrees, expressing a valid bounding box. The maximal width (\code{right - left}) and height (\code{top - bottom}) of the
bounding box is 0.25 degree.}

\item{page_number}{Specifies which group of 5,000 points, or page, to return. Since the command does not return more
than 5,000 points at a time, this parameter must be incremented —and the command sent again (using the same bounding
box)— in order to retrieve all of the points for a bounding box that contains more than 5,000 points. When this
parameter is 0 (zero), the command returns the first 5,000 points; when it is 1, the command returns points
5,001–10,000, etc.}
}
\description{
Use this to retrieve the GPS track points that are inside a given bounding box (formatted in a GPX format).
}
\note{
In violation of the \href{https://www.topografix.com/GPX/1/1/#type_trksegType}{GPX standard} when downloading public
GPX traces through the API, all waypoints of non-trackable traces are randomized (or rather sorted by lat/lon) and
delivered as one trackSegment for privacy reasons. Trackable traces are delivered, sorted by descending upload
time, before the waypoints of non-trackable traces.
}
\examples{
pts_gps <- osm_get_points_gps(bbox = c(-0.3667545, 40.2153246, -0.3354263, 40.2364915))
## bbox as a character value also works. Equivalent call:
# pts_gps <- osm_get_points_gps(bbox = "-0.3667545,40.2153246,-0.3354263,40.2364915")
pts_gps
}
\seealso{
Other GPS' functions: 
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_list_gpxs}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{GPS' functions}
