% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_user_data.R
\name{osm_get_preferences_user}
\alias{osm_get_preferences_user}
\title{Preferences of the logged-in user}
\usage{
osm_get_preferences_user(key, format = c("R", "xml", "json"))
}
\arguments{
\item{key}{Returns a string with that preference's value. If missing, return all preferences.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}. Only relevant when \code{key} is missing.}
}
\value{
If \code{format = "R"}, returns a data frame with \code{key} and \code{value} columns of the user preferences.
\subsection{\code{format = "xml"}}{

Returns a \link[xml2:oldclass]{xml2::xml_document} with the following format:

\if{html}{\out{<div class="sourceCode xml">}}\preformatted{<osm version="0.6" generator="OpenStreetMap server">
	<preferences>
		<preference k="somekey" v="somevalue" />
		...
	</preferences>
</osm>
}\if{html}{\out{</div>}}
}

\subsection{\code{format = "json"}}{

Returns a list with the following json structure:

\if{html}{\out{<div class="sourceCode json">}}\preformatted{\{
 "version": "0.6",
 "generator": "OpenStreetMap server",
 "preferences": \{"somekey": "somevalue, ...\}
\}
}\if{html}{\out{</div>}}
}
}
\description{
Preferences of the logged-in user
}
\examples{
\dontrun{
prefs <- osm_get_preferences_user()
prefs
}
}
\seealso{
Other users' functions: 
\code{\link{osm_details_logged_user}()},
\code{\link{osm_get_user_details}()}
}
\concept{users' functions}
