% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_get_user_details.R
\name{osm_get_user_details}
\alias{osm_get_user_details}
\title{Details of users}
\usage{
osm_get_user_details(user_id, format = c("R", "xml", "json"))
}
\arguments{
\item{user_id}{The ids of the users to retrieve the details for, represented by a numeric or a character value (not
the display names).}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}
}
\description{
Details of users
}
\examples{
\dontrun{
usrs <- osm_details_users(user_ids = c(1, 24, 44, 45, 46, 48, 49, 50))
usrs
}
}
\seealso{
Other users' functions: 
\code{\link{osm_details_logged_user}()},
\code{\link{osm_get_preferences_user}()}
}
\concept{users' functions}
