% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_traces.R
\name{osm_list_gpxs}
\alias{osm_list_gpxs}
\title{List user's GPX traces}
\usage{
osm_list_gpxs()
}
\description{
Use this to get a list of GPX traces owned by the authenticated user. Requires authentication.
}
\examples{
\dontrun{
traces <- osm_list_gpxs()
traces
}
}
\seealso{
Other get GPS' functions: 
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()}
}
\concept{get GPS' functions}
