% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changesets.R
\name{osm_read_changeset}
\alias{osm_read_changeset}
\title{Read a changeset}
\usage{
osm_read_changeset(changeset_id, include_discussion = FALSE)
}
\arguments{
\item{changeset_id}{The id of the changeset to retrieve represented by a numeric or a character value.}

\item{include_discussion}{Indicates whether the result should contain the changeset discussion or not.}
}
\description{
Returns the changeset with the given \code{changeset_id}.
}
\details{
\itemize{
\item The \code{uid} might not be available for changesets auto generated by the API v0.5 to API v0.6 transition
\item The bounding box attributes will be missing for an empty changeset.
\item The changeset bounding box is a rectangle that contains the bounding boxes of all objects changed in this
changeset. It is not necessarily the smallest possible rectangle that does so.
\item This API call only returns information about the changeset itself but not the actual changes made to elements in
this changeset. To access this information use \code{\link[=osm_download_changeset]{osm_download_changeset()}}.
}
}
\examples{
chaset <- osm_read_changeset(changeset_id = 137595351, include_discussion = TRUE)
chaset
}
\seealso{
Other changesets' functions: 
\code{\link{osm_download_changeset}()},
\code{\link{osm_query_changesets}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{changesets' functions}
