% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_read_note}
\alias{osm_read_note}
\title{Read notes}
\usage{
osm_read_note(note_id, format = c("R", "xml", "rss", "json", "gpx"))
}
\arguments{
\item{note_id}{Note id represented by a numeric or a character value.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, \code{rss}, \code{json} or \code{gpx}.}
}
\description{
Returns the existing note with the given ID.
}
\examples{
\dontrun{
note <- osm_read_note(note_id = "2067786")
note
}
}
\seealso{
Other get notes' functions: 
\code{\link{osm_feed_notes}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_search_notes}()}
}
\concept{get notes' functions}
