% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_reopen_note}
\alias{osm_reopen_note}
\title{Reopen a note}
\usage{
osm_reopen_note(note_id)
}
\arguments{
\item{note_id}{Note id represented by a numeric or a character value.}
}
\description{
Reopen a closed note. Requires authentication.
}
\seealso{
Other edit notes' functions: 
\code{\link{osm_close_note}()},
\code{\link{osm_create_comment_note}()},
\code{\link{osm_create_note}()}
}
\concept{edit notes' functions}
