% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_search_notes}
\alias{osm_search_notes}
\title{Search for notes}
\usage{
osm_search_notes(
  q,
  user,
  from,
  to,
  closed = 7,
  sort = c("updated_at", "created_at"),
  order = c("newest", "oldest"),
  limit = 100,
  format = c("data.frame", "xml", "rss", "json", "gpx")
)
}
\arguments{
\item{q}{Specifies the search query.}

\item{user}{Find notes by the user with the given user id (numeric) or display name (character).}

\item{from}{Specifies the beginning of a date range to search in for a note. A valid
\href{https://en.wikipedia.org/wiki/ISO_8601}{ISO 8601} date.}

\item{to}{Specifies the end of a date range to search in for a note. A valid
\href{https://en.wikipedia.org/wiki/ISO_8601}{ISO 8601} date. The date of today is the default.}

\item{closed}{Specifies the number of days a note needs to be closed to no longer be returned. A value of 0 means
only open notes are returned. A value of -1 means all notes are returned. 7 is the default.}

\item{sort}{Specifies the value which should be used to sort the notes. It is either possible to sort them by their
creation date (\code{created_at}) or the date of the last update (\code{updated_at}, the default).}

\item{order}{Specifies the order of the returned notes. It is possible to order them in ascending (\code{oldest}) or
descending (\code{newest}, the default) order.}

\item{limit}{Specifies the number of entries returned at max. A value of between 1 and 10000 is valid.    100 is the
default.}

\item{format}{Format of the the returned list of notes. Can be \code{data.frame} (default), \code{xml}, \code{rss}, \code{json} or \code{gpx}.}
}
\description{
Returns the existing notes matching either the initial note text or any of the comments.
}
\details{
The notes will be ordered by the date of their last change, the most recent one will be first. If no query was
specified, the latest notes are returned.
}
\examples{
notes <- osm_search_notes(q = "POI", from = "2017-10-01", to = "2017-10-27T15:27A", limit = 10)
notes

my_notes <- osm_search_notes(user = "jmaspons", closed = -1, format = "json")
my_notes
}
\seealso{
Other notes' functions: 
\code{\link{osm_feed_notes}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_note}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{notes' functions}
