% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_map_notes.R
\name{osm_subscribe_note}
\alias{osm_subscribe_note}
\alias{osm_unsubscribe_note}
\title{Subscribe or unsubscribe to a note}
\usage{
osm_subscribe_note(note_id)

osm_unsubscribe_note(note_id)
}
\arguments{
\item{note_id}{The id of the note represented by a numeric or a character value.}
}
\value{
Returns nothing.
}
\description{
Subscribe or unsubscribe to a note
}
\section{Functions}{
\itemize{
\item \code{osm_subscribe_note()}: Subscribe to the discussion of a note to receive notifications
for new comments.

\item \code{osm_unsubscribe_note()}: Unsubscribe from the discussion of a note to stop receiving notifications for new
comments.

}}
\examples{
\dontrun{
# set_osmapi_connection(server = "openstreetmap.org")
osm_subscribe_note(2067786)
osm_unsubscribe_note("2067786")
}
}
\concept{subscription to notes' functions}
