% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeset_discussion.R
\name{osm_unsubscribe_changeset_discussion}
\alias{osm_unsubscribe_changeset_discussion}
\title{Unsubscribe from a changeset discussion}
\usage{
osm_unsubscribe_changeset_discussion(changeset_id)
}
\arguments{
\item{changeset_id}{The id of the changeset to unsubscribe to represented by a numeric or a character value.}
}
\value{
Returns the changeset information invisibly as a \code{xml_document}.
}
\description{
Unsubscribe from the discussion of a changeset to stop receiving notifications.
}
\seealso{
Other changeset discussion's functions: 
\code{\link{osm_comment_changeset_discussion}()},
\code{\link{osm_subscribe_changeset_discussion}()}
}
\concept{changeset discussion's functions}
