% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_osmapi_url}
\alias{get_osmapi_url}
\title{Get current OSM API URL}
\usage{
get_osmapi_url()
}
\value{
the OSM API URL currently in use.
}
\description{
Return the URL of the currently specified OSM API. Default is \url{https://api.openstreetmap.org}.
}
\seealso{
Other API functions: 
\code{\link{osm_api_versions}()},
\code{\link{osm_capabilities}()},
\code{\link{set_osmapi_url}()}
}
\concept{API functions}
