% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_miscellaneous.R
\name{osm_api_versions}
\alias{osm_api_versions}
\title{Available API versions}
\usage{
osm_api_versions()
}
\value{
A character vector with the supported versions
}
\description{
Available API versions
}
\examples{
osm_api_versions()
}
\seealso{
Other API functions: 
\code{\link{authenticate_osmapi}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_permissions}()},
\code{\link{set_osmapi_connection}()}
}
\concept{API functions}
