% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{osm_capabilities}
\alias{osm_capabilities}
\title{Capabilities of the API}
\usage{
osm_capabilities()
}
\value{
A list with the API capabilities and policies.
}
\description{
Provide information about the capabilities and limitations of the current API.
}
\details{
API:
\itemize{
\item Version minimum and maximum are the API call versions that the server will accept.
\item Area maximum is the maximum area in square degrees that can be queried by API calls.
\item Tracepoints per_page is the maximum number of points in a single GPS trace. (Possibly incorrect)
\item Waypoints maximum is the maximum number of nodes that a way may contain.
\item Relation member maximum is the maximum number of members that a relation may contain.
\item Changesets maximum is the maximum number of combined nodes, ways and relations that can be contained in a
changeset.
\item The status element returns either ''online'', ''readonly'' or ''offline'' for each of the database, API and GPX
API. The database field is informational, and the API/GPX-API fields indicate whether a client should expect read
and write requests to work (''online''), only read requests to work (''readonly'') or no requests to work
(''offline'').
}

Policy:
\itemize{
\item Imagery blacklist lists all aerial and map sources, which are not permitted for OSM usage due to copyright. Editors
must not show these resources as background layer.
}
}
\examples{
osm_capabilities()
}
\seealso{
Other API functions: 
\code{\link{authenticate_osmapi}()},
\code{\link{osm_api_versions}()},
\code{\link{osm_permissions}()},
\code{\link{set_osmapi_connection}()}
}
\concept{API functions}
