% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_close_note}
\alias{osm_close_note}
\title{Close a note}
\usage{
osm_close_note(note_id)
}
\arguments{
\item{note_id}{Note id represented by a numeric or a character value.}
}
\description{
Close a note as fixed. Requires authentication.
}
\seealso{
Other edit notes' functions: 
\code{\link{osm_create_comment_note}()},
\code{\link{osm_create_note}()},
\code{\link{osm_reopen_note}()}
}
\concept{edit notes' functions}
