% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_gps_traces.R
\name{osm_delete_gpx}
\alias{osm_delete_gpx}
\title{Delete GPS trace}
\usage{
osm_delete_gpx(gpx_id)
}
\arguments{
\item{gpx_id}{The track id represented by a numeric or a character value.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Use this to delete a GPX file. Only usable by the owner account. Requires authentication.
}
\examples{
vignette("how_to_edit_gps_traces", package = "osmapiR")
}
\seealso{
Other edit GPS traces' functions: 
\code{\link{osm_create_gpx}()},
\code{\link{osm_update_gpx}()}
}
\concept{edit GPS traces' functions}
