% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_delete_gpx.R
\name{osm_delete_gpx}
\alias{osm_delete_gpx}
\title{Delete GPS traces}
\usage{
osm_delete_gpx(gpx_id)
}
\arguments{
\item{gpx_id}{The track ids represented by a numeric or a character vector.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Use this to delete GPX files. Only usable by the owner account. Requires authentication.
}
\examples{
vignette("how_to_edit_gps_traces", package = "osmapiR")
}
\seealso{
Other edit GPS traces' functions: 
\code{\link{osm_create_gpx}()},
\code{\link{osm_update_gpx}()}
}
\concept{edit GPS traces' functions}
