% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_user_data.R
\name{osm_details_logged_user}
\alias{osm_details_logged_user}
\title{Details of the logged-in user}
\usage{
osm_details_logged_user(format = c("R", "xml", "json"))
}
\arguments{
\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}
}
\value{
A list with the user details.
}
\description{
You can get the home location, the display name of the user and other details.
}
\examples{
\dontrun{
usr_details <- osm_details_logged_user()
usr_details
}
}
\seealso{
Other users' functions: 
\code{\link{osm_get_preferences_user}()},
\code{\link{osm_get_user_details}()}
}
\concept{users' functions}
