% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data.R
\name{osm_details_user}
\alias{osm_details_user}
\title{Details of a user}
\usage{
osm_details_user(user_id)
}
\arguments{
\item{user_id}{The id of the user to retrieve represented by a numeric or a character value (not the display name).}
}
\description{
Details of a user
}
\examples{
usr <- osm_details_user(user_id = "11725140")
usr
}
\seealso{
Other users' functions: 
\code{\link{osm_details_users}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_users}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{users' functions}
