% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_notes.R
\name{osm_feed_notes}
\alias{osm_feed_notes}
\title{RSS Feed of notes in a bbox}
\usage{
osm_feed_notes(bbox)
}
\arguments{
\item{bbox}{Coordinates for the area to retrieve the notes from (\verb{left,bottom,right,top}). Floating point numbers in
degrees, expressing a valid bounding box.}
}
\description{
RSS Feed of notes in a bbox
}
\examples{
feed_notes <- osm_feed_notes(bbox = c(0.8205414, 40.6686604, 0.8857727, 40.7493377))
## bbox as a character value also works. Equivalent call:
# feed_notes <- osm_feed_notes(bbox = "0.8205414,40.6686604,0.8857727,40.7493377")
feed_notes
}
\seealso{
Other notes' functions: 
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_note}()},
\code{\link{osm_search_notes}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{notes' functions}
