% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_traces.R
\name{osm_get_data_gpx}
\alias{osm_get_data_gpx}
\title{Download GPS Track Data}
\usage{
osm_get_data_gpx(gpx_id, format)
}
\arguments{
\item{gpx_id}{The track id represented by a numeric or a character value.}

\item{format}{Format of the output. If missing (default), the response will be the exact file that was uploaded.
If \code{R}, a \code{data.frame}.
If \code{gpx}, the response will always be a GPX format file.
If \code{xml}, a \code{XML} file in an undocumented format.}
}
\description{
Use this to download the full GPX file. Available without authentication if the file is marked public. Otherwise only
usable by the owner account and requires authentication.
}
\note{
If you request refers to a multi-file archive the response when you force gpx or xml format will consist of a
non-standard simple concatenation of the files.
}
\examples{
\dontrun{
trk_data <- osm_get_data_gpx(gpx_id = 3498170, format = "R")
trk_data
}
}
\seealso{
Other GPS' functions: 
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_list_gpxs}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_get_preferences_user}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{GPS' functions}
