% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_user_data.R
\name{osm_get_preferences_user}
\alias{osm_get_preferences_user}
\title{Preferences of the logged-in user}
\usage{
osm_get_preferences_user(key, format = c("R", "xml", "json"))
}
\arguments{
\item{key}{Returns a string with that preference's value. If missing, return all preferences.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}. Only relevant when \code{key} is missing.}
}
\description{
Preferences of the logged-in user
}
\examples{
\dontrun{
prefs <- osm_get_preferences_user()
prefs
}
}
\seealso{
Other users' functions: 
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()}
}
\concept{users' functions}
