% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeset_discussion.R
\name{osm_hide_comment_changeset_discussion}
\alias{osm_hide_comment_changeset_discussion}
\title{Hide changeset comment}
\usage{
osm_hide_comment_changeset_discussion(comment_id)
}
\arguments{
\item{comment_id}{Note that the changeset comment id differs from the changeset id.}
}
\description{
Sets visible flag on changeset comment to false. This request needs to be done as an authenticated user with
moderator role.
}
\seealso{
Other functions for moderators: 
\code{\link{osm_redaction_object}()},
\code{\link{osm_unhide_comment_changeset_discussion}()}
}
\concept{functions for moderators}
