% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_traces.R
\name{osm_list_gpxs}
\alias{osm_list_gpxs}
\title{List user's GPX traces}
\usage{
osm_list_gpxs()
}
\description{
Use this to get a list of GPX traces owned by the authenticated user. Requires authentication.
}
\examples{
\dontrun{
traces <- osm_list_gpxs()
traces
}
}
\seealso{
Other GPS' functions: 
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_history_object}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{GPS' functions}
