% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changesets.R
\name{osm_query_changesets}
\alias{osm_query_changesets}
\title{Query changesets}
\usage{
osm_query_changesets(
  bbox,
  user,
  time,
  time_2,
  open,
  closed,
  changeset_ids,
  limit = 100,
  format = c("R", "xml", "json"),
  tags_in_columns = FALSE
)
}
\arguments{
\item{bbox}{Find changesets within the given bounding box coordinates (\verb{left,bottom,right,top}).}

\item{user}{Find changesets by the user with the given user id (numeric) or display name (character).}

\item{time}{Find changesets \strong{closed} after this date and time. See details for the valid formats.}

\item{time_2}{find changesets that were \strong{closed} after \code{time} and \strong{created} before \code{time_2}. In other words, any
changesets that were open \strong{at some time} during the given time range \code{time} to \code{time_2}.}

\item{open}{If \code{TRUE}, only finds changesets that are still \strong{open} but excludes changesets that are closed or have
reached the element limit for a changeset (10,000 at the moment \code{osm_capabilities()$api$changesets}).}

\item{closed}{If \code{TRUE}, only finds changesets that are \strong{closed} or have reached the element limit.}

\item{changeset_ids}{Finds changesets with the specified ids.}

\item{limit}{Specifies the maximum number of changesets returned. A number between 1 and 100, with 100 as the default
value.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}

\item{tags_in_columns}{If \code{FALSE} (default), the tags of the changesets are saved in a single list column \code{tags}
containing a \code{data.frame} for each changeset with the keys and values. If \code{TRUE}, add a column for each key.
Ignored if \code{format != "R"}.}
}
\description{
This is an API method for querying changesets. It supports querying by different criteria.
}
\details{
Where multiple queries are given the result will be those which match all of the requirements. The contents of the
returned document are the changesets and their tags. To get the full set of changes associated with a changeset, use
\code{\link[=osm_download_changeset]{osm_download_changeset()}} on each changeset ID individually.

Modification and extension of the basic queries above may be required to support rollback and other uses we find for
changesets.

This call returns latest changesets matching criteria,
\href{https://github.com/openstreetmap/openstreetmap-website/blob/f1c6a87aa137c11d0aff5a4b0e563ac2c2a8f82d/app/controllers/api/changesets_controller.rb#L174}{ordered by \code{created_at}}
– see the \href{https://github.com/openstreetmap/openstreetmap-website/blob/master/app/controllers/api/changesets_controller.rb#L174}{current state}.

Te valid formats for \code{time} and \code{time_2} parameters are anything that
\href{https://ruby-doc.org/stdlib-2.7.0/libdoc/time/rdoc/Time.html#method-c-parse}{\code{Time.parse} Ruby function} will parse.
}
\examples{
\dontrun{
chst_ids <- osm_query_changesets(changeset_ids = c(137627129, 137625624))
chst_ids

chsts <- osm_query_changesets(
  bbox = c(-1.241112, 38.0294955, 8.4203171, 42.9186456),
  user = "Mementomoristultus",
  time = "2023-06-22T02:23:23Z",
  time_2 = "2023-06-22T00:38:20Z"
)
chsts

chsts2 <- osm_query_changesets(
  bbox = c("-9.3015367,41.8073642,-6.7339533,43.790422"),
  user = "Mementomoristultus",
  closed = TRUE
)
chsts2
}
}
\seealso{
Other get changesets' functions: 
\code{\link{osm_download_changeset}()},
\code{\link{osm_read_changeset}()}
}
\concept{get changesets' functions}
