% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_map_notes.R
\name{osm_read_bbox_notes}
\alias{osm_read_bbox_notes}
\title{Retrieve notes by bounding box}
\usage{
osm_read_bbox_notes(
  bbox,
  limit = 100,
  closed = 7,
  format = c("R", "xml", "rss", "json", "gpx")
)
}
\arguments{
\item{bbox}{Coordinates for the area to retrieve the notes from (\verb{left,bottom,right,top}). Floating point numbers in
degrees, expressing a valid bounding box, not larger than the configured size limit, 25 square degrees, not
overlapping the dateline.}

\item{limit}{Specifies the number of entries returned at max. A value between 1 and 10000 is valid. Default to 100.}

\item{closed}{Specifies the number of days a note needs to be closed to no longer be returned. A value of 0 means
only open notes are returned. A value of -1 means all notes are returned. Default to 7.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, \code{rss}, \code{json} or \code{gpx}.}
}
\description{
Returns the existing notes in the specified bounding box. The notes will be ordered by the date of their last change,
the most recent one will be first.
}
\note{
The comment properties (\code{uid}, \code{user}, \code{user_url}) will be omitted if the comment was anonymous.
}
\examples{
notes <- osm_read_bbox_notes(bbox = c(3.7854767, 39.7837403, 4.3347931, 40.1011851), limit = 10)
## bbox as a character value also works. Equivalent call:
# osm_read_bbox_notes(bbox = "3.7854767,39.7837403,4.3347931,40.1011851", limit = 10)
notes
}
\seealso{
Other get notes' functions: 
\code{\link{osm_feed_notes}()},
\code{\link{osm_read_note}()},
\code{\link{osm_search_notes}()}
}
\concept{get notes' functions}
