% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_elements.R
\name{osm_redaction_object}
\alias{osm_redaction_object}
\title{Redact an object version}
\usage{
osm_redaction_object(
  osm_type = c("node", "way", "relation"),
  osm_id,
  version,
  redaction_id
)
}
\arguments{
\item{osm_type}{Object type (\code{"node"}, \code{"way"} or \code{"relation"}).}

\item{osm_id}{Object id represented by a numeric or a character value.}

\item{version}{Version of the object to redact.}

\item{redaction_id}{If missing, then this is an unredact operation. If a redaction ID was specified, then set this
element to be redacted in that redaction.}
}
\description{
Used by the \href{https://wiki.openstreetmap.org/wiki/Data_working_group}{Data Working Group} to hide old versions of
elements containing data privacy or copyright infringements. Only permitted for OSM accounts with the moderator role
(DWG and server admins).
}
\details{
The \code{redaction_id} is listed on \url{https://www.openstreetmap.org/redactions}. More information can be found in
\href{https://git.openstreetmap.org/rails.git/blob/HEAD:/app/controllers/api/old_controller.rb}{the source}.
}
\seealso{
Other functions for moderators: 
\code{\link{osm_delete_note}()},
\code{\link{osm_hide_comment_changeset_discussion}()},
\code{\link{osm_unhide_comment_changeset_discussion}()}
}
\concept{functions for moderators}
