% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{osm_relations_object}
\alias{osm_relations_object}
\title{Relations of an object}
\usage{
osm_relations_object(
  osm_type = c("node", "way", "relation"),
  osm_id,
  format = c("R", "xml", "json"),
  tags_in_columns = FALSE
)
}
\arguments{
\item{osm_type}{Object type (\code{"node"}, \code{"way"} or \code{"relation"}).}

\item{osm_id}{Object id represented by a numeric or a character value.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}

\item{tags_in_columns}{if \code{FALSE} (default), the tags of the objects are saved in a single list column \code{tags}
containing a \code{data.frame} for each OSM object with the keys and values. If \code{TRUE}, add a column for each key.
Ignored if \code{format != "R"}.}
}
\description{
Returns all (not deleted) relations in which the given object is used.
}
\examples{
node <- osm_relations_object(osm_type = "node", osm_id = 152364165)
node

way <- osm_relations_object(osm_type = "way", osm_id = 372011578)
way

rel <- osm_relations_object(osm_type = "relation", osm_id = 342792)
rel
}
\seealso{
Other OSM objects' functions: 
\code{\link{osm_bbox_objects}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_history_object}()},
\code{\link{osm_read_object}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}

Other GET calls: 
\code{\link{osm_api_versions}()},
\code{\link{osm_bbox_objects}()},
\code{\link{osm_capabilities}()},
\code{\link{osm_details_logged_user}()},
\code{\link{osm_details_users}()},
\code{\link{osm_details_user}()},
\code{\link{osm_download_changeset}()},
\code{\link{osm_feed_notes}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_get_data_gpx}()},
\code{\link{osm_get_metadata_gpx}()},
\code{\link{osm_get_points_gps}()},
\code{\link{osm_get_preferences_user}()},
\code{\link{osm_history_object}()},
\code{\link{osm_list_gpxs}()},
\code{\link{osm_permissions}()},
\code{\link{osm_query_changesets}()},
\code{\link{osm_read_bbox_notes}()},
\code{\link{osm_read_changeset}()},
\code{\link{osm_read_note}()},
\code{\link{osm_read_object}()},
\code{\link{osm_search_notes}()},
\code{\link{osm_version_object}()},
\code{\link{osm_ways_node}()}
}
\concept{GET calls}
\concept{OSM objects' functions}
