% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_traces.R
\name{osm_update_gpx}
\alias{osm_update_gpx}
\title{Update GPS trace}
\usage{
osm_update_gpx(gpx_id, file)
}
\arguments{
\item{gpx_id}{The track id represented by a numeric or a character value.}

\item{file}{The GPX file path containing the track points.}
}
\description{
Use this to update a GPX file. Only usable by the owner account. Requires authentication.
}
