% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapi_elements.R
\name{osm_ways_node}
\alias{osm_ways_node}
\title{Ways of a node}
\usage{
osm_ways_node(node_id, format = c("R", "xml", "json"), tags_in_columns = FALSE)
}
\arguments{
\item{node_id}{Node id represented by a numeric or a character value.}

\item{format}{Format of the output. Can be \code{R} (default), \code{xml}, or \code{json}.}

\item{tags_in_columns}{If \code{FALSE} (default), the tags of the objects are saved in a single list column \code{tags}
containing a \code{data.frame} for each OSM object with the keys and values. If \code{TRUE}, add a column for each key.
Ignored if \code{format != "R"}.}
}
\description{
Returns all the (not deleted) ways in which the given node is used.
}
\examples{
ways_node <- osm_ways_node(node_id = 35308286)
ways_node
}
\seealso{
Other get OSM objects' functions: 
\code{\link{osm_bbox_objects}()},
\code{\link{osm_fetch_objects}()},
\code{\link{osm_full_object}()},
\code{\link{osm_history_object}()},
\code{\link{osm_read_object}()},
\code{\link{osm_relations_object}()},
\code{\link{osm_version_object}()}
}
\concept{get OSM objects' functions}
