% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmapiR-package.R
\docType{package}
\name{osmapiR-package}
\alias{osmapiR}
\alias{osmapiR-package}
\title{osmapiR: OpenStreetMap API}
\description{
Implements OpenStreetMap API v0.6 (\url{https://wiki.openstreetmap.org/wiki/API_v0.6}).
}
\details{
You are responsible for following the \href{https://operations.osmfoundation.org/policies/api/}{API Usage Policy}.

You can modify the user agent of the requests by setting the option \code{osmapir.user_agent}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(osmapir.user_agent = "my new user agent")
}\if{html}{\out{</div>}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmaspons.github.io/osmapiR https://github.com/jmaspons/osmapiR}
  \item Report bugs at \url{https://github.com/jmaspons/osmapiR/issues}
}

}
\author{
\strong{Maintainer}: Joan Maspons \email{joanmaspons@gmail.com} (\href{https://orcid.org/0000-0003-2286-8727}{ORCID})

}
\keyword{internal}
