% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmchange.R
\name{osmchange_create}
\alias{osmchange_create}
\title{\code{osmchange} to create OSM objects}
\usage{
osmchange_create(x)
}
\arguments{
\item{x}{A \link{osmapi_objects} with columns \code{type}, \code{changeset} + column \code{members} for ways and relations + \code{lat}
and \code{lon} for nodes + tags if needed.}
}
\value{
Returns a \code{osmapi_OsmChange} data frame with one OSM object per row ready to send the editions to the servers
with \code{\link[=osm_diff_upload_changeset]{osm_diff_upload_changeset()}}.
}
\description{
Prepare data to create OSM objects.
}
\details{
Objects IDs are unknown and will be allocated by the server. Check
\href{https://wiki.openstreetmap.org/wiki/OsmChange}{OsmChange page} for details about how to refer to objects still not
created to define the members of relations and nodes of ways.
}
\seealso{
Other OsmChange's functions: 
\code{\link{osm_diff_upload_changeset}()},
\code{\link{osm_download_changeset}()},
\code{\link{osmchange_delete}()},
\code{\link{osmchange_modify}()}
}
\concept{OsmChange's functions}
