% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmchange.R
\name{osmchange_create}
\alias{osmchange_create}
\title{Create OSM objects}
\usage{
osmchange_create(x)
}
\arguments{
\item{x}{A \code{data.frame} with columns \code{type}, \code{changeset} and \code{tags} + column \code{members} for ways and relations + \code{lat}
and \code{lon} for nodes. For \code{osmapi_obects}, the tags column is not needed but the object must inherit \code{tags_wide}.}
}
\description{
Create OSM objects
}
\details{
Objects IDs are unknown and will be allocated by the server. Check
\href{https://wiki.openstreetmap.org/wiki/OsmChange}{OsmChange page} for details about how to refer to objects still not
created to define the members of relations and nodes of ways.
}
\seealso{
Other OsmChange's functions: 
\code{\link{osm_diff_upload_changeset}()},
\code{\link{osm_download_changeset}()},
\code{\link{osmchange_delete}()},
\code{\link{osmchange_modify}()}
}
\concept{OsmChange's functions}
