% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_sf.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.osmapi_map_notes}
\title{Convert osmapiR objects to sf objects}
\usage{
st_as_sf.osmapi_map_notes(x, ...)
}
\arguments{
\item{x}{an osmapiR object.}

\item{...}{passed on to \code{st_sf} form \pkg{sf}.}
}
\value{
Returns a \code{sf} object from \pkg{sf}.
}
\description{
Convert osmapiR objects to sf objects
}
\seealso{
Other methods: 
\code{\link{tags_list2wide}()}
}
\concept{methods}
